Refactor set op record batch and probe template
